requirements.txt
bsddb3>=*
pycryptodome>=*
ecdsa>=*

# dev tools
pytest>=*
